meta-qti-automotive: add scm_user_intf_sec.ko to sa8797p early ramdisk
Add scm_user_intf_sec kernel module to early ramdisk kernel modules, so
it will be installed on rootfs folder.
After it will be installed in the rootfs folder we will be able to load
it early in the boot process using early-ramdisk-init.

We need this since smcinvoke is loaded early in the boot, but it relies
on scm_user_intf_sec module to be loaded already, so we need to make
sure it's loaded earlier, before smcinvoke is loaded.

Change-Id: I12cf997d28a73e0281548e99e8d307a0988ac205
Signed-off-by: Yiftach Ariel <quic_yariel@quicinc.com>