meta-qti-automotive: Add tz_ffi.ko to early ramdisk
Add the tz_ffi kernel module to early ramdisk kernel modules, so it will
be installed on rootfs folder.
After it will be installed in the rootfs folder we will be able to load
it early in the boot process using early-ramdisk-init.

We need the tz_ffi kernel module to be loaded early before any call to
TrustZone because it tells TrustZone which HLOS memory regions can be
set as RW (other memory regions will be defaulted to RO).

Change-Id: If99b6a72efb2ee22615d8044a95758e390aac5b9
Signed-off-by: Yiftach Ariel <yariel@qti.qualcomm.com>