meta-qti-automotive: Update dtb recipe for ddr size based dt overlay
In the device tree, certain nodes need to be updated based on the
DDR size available on the target. To achieve this, we can overlay
the device tree (dt) on top of dtb files generated for different
boards. We introduce OOT_DDR_DTBOS variable for the overlay device
tree created for various DDR sizes.

The device tree is selected by ABL based on the DDR type defined
in the board-id subtype value. Bits 8-10 in the board-id subtype
field are reserved for DDR type. This subtype value is updated
during the overlay process.

Below value are used for differnt ddr type
  DDRTYPE_8GB = 1
  DDRTYPE_12GB = 2
  DDRTYPE_16GB = 3
  DDRTYPE_24GB = 4
  DDRTYPE_36GB = 5
  DDRTYPE_48GB = 6
  DDRTYPE_64GB = 7

Change-Id: I03334a6a3b17a04fd57f3aec42f0e2445c5e20ab
Signed-off-by: Jai Kumar Gautam <jgautam@qti.qualcomm.com>