meta-qti-automotive: Exit with code 0 to suppress service restarts
Currently, the eth1 interface is not present, which causes the
script to return exit code 1 after attempting to check the interface.
This, in turn, triggers repeated restarts of the setup_eth1.service.

To avoid this issue, we’ve changed the failure exit code from 1 to 0.
The eth1 interface may exist in other projects, making it difficult
to distinguish whether it's existed or not.
Additionally, disabling the restart behavior at the service level
is not a viable solution, as service failures may not be permitted
in our environment.

Reduce the number of attempts and the sleep duration.

Change-Id: Ib75373b3de29b2fcd6a48e8ca0900b77dc783b78
Signed-off-by: Mei Xie <quic_meixie@quicinc.com>