systemd-conf: modified RebootWatchdogSec value
The watchdog clock rate is now correctly set to 32765 Hz. At this rate, a 32-second timeout results in a tick count of 0x100000, which exceeds the 20-bit width of the bite register and gets truncated to 0—effectively disabling the watchdog.

To ensure the timeout value fits within the register and functions as expected, RebootWatchdogSec is reduced to 30 seconds. This results in a tick count of 0xffd2d, which is safely within the 20-bit limit.

Change-Id: I35d09910fd3e11ed466fa14ac05bed11d16ae27b
Signed-off-by: Venkatakrishanaiah Pari <vpari@qti.qualcomm.com>