meta-qti-automotive: move machine conf contents to gen5.conf
To support unified build, we need use gen5.conf for all gen5 targets.
However, before the migration work is fully completed, it is still
necessary to keep sa8797.conf as a transitional measure. For now, all
the contents of sa8797.conf will be moved into gen5.conf, and
sa8797.conf will simply serve as a wrapper for gen5.conf. Once the
migration is complete, sa8797.conf will be removed.

Signed-off-by: Naiwen Yu <naiwyu@qti.qualcomm.com>
Change-Id: I996927f87b2d406ba67690e31f4506f25f8d116f