meta-qti-automotive: Update oot-dtbo recipe to support ddr overlay
In the device tree, certain nodes need to be updated based on the
DDR size available on the target. To achieve this, we can overlay
the device tree (dt) on top of dtb files generated for different
boards. We introduce OOT_DDR_DTBOS variable for the overlay device
tree created for various DDR sizes.

Change-Id: I810721fd90830c84cc660a17a5201055ba67bd61
Signed-off-by: hanzh <hanzh@qti.qualcomm.com>